+++ src/index.js
@@ -3,7 +3,6 @@
 const core = require("@actions/core");
 // const github = require("@actions/github");
 const isbn = require("node-isbn");
-const fetch = require("node-fetch");
 const { titleParser, getBook, writeFile } = require("./utils");
 
 async function read() {
@@ -16,25 +15,13 @@ async function read() {
     const providers = core.getInput("providers")
       ? core.getInput("providers").split(",")
       : isbn._providers;
-    const imageDirectory = core.getInput("imageDirectory");
     core.exportVariable("IssueNumber", number);
     const bookMetadata = await getBook(
       { date, body, bookIsbn, providers },
       fileName
     );
-    core.warning(bookMetadata);
     // Write book to yaml file
     await writeFile(fileName, bookMetadata);
-    // Download book thumbnail
-    if (bookMetadata.imageLinks.thumbnail) {
-      try {
-        const response = await fetch(bookMetadata.imageLinks.thumbnail);
-        const buffer = await response.buffer();
-        await writeFile(`${imageDirectory}/book-${bookIsbn}.png`, buffer);
-      } catch (error) {
-        core.setFailed(error.err);
-      }
-    }
   } catch (error) {
     core.setFailed(error.message);
   }
